+++ components/navbar/AccountSection.vue
@@ -2,7 +2,11 @@
   <div>
     <LoginButton v-if="!isAuthenticated" />
     <div v-else class="logged-user">
-      <avatar class="avatar" @click="redirectToUser" />
+      <avatar
+        class="avatar"
+        :avatar-url="me.avatarUrl"
+        @click="redirectToUser"
+      />
       <span class="user--name" @click="redirectToUser">{{
         me.name || me.login
       }}</span>
@@ -41,6 +45,9 @@ export default {
 }
 </script>
 <style lang="scss" scoped>
+.user--name:hover {
+  cursor: pointer;
+}
 .avatar {
   width: 40px;
   height: 40px;